requirements.txt
# influxdb-client
influxdb
pyserial
RPi.GPIO
pyyaml
gpiozero
crate